chore: exports PlotObject for correct type inference
